only copy file if not in workspace

+++ entrypoint.sh
@@ -96,5 +96,7 @@ terraform ${TF_VERB} ${TF_PLAN} ${TF_VARSFILE} ${TF_AUTOAPPROVE} ${TF_OUT}
 # Copy tfplan to github workspace
 if [ "$TF_VERB" = "plan" ]
 then
-    cp tfplan /github/workspace/$INPUT_PLANFILE
+    if [ ! -f /github/workspace/$INPUT_PLANFILE ] then
+      cp tfplan /github/workspace/$INPUT_PLANFILE
+    fi
 fi
